Restructures cli files
