+++ aqueduct/lib/src/cli/commands/db_validate.dart
@@ -1,12 +1,10 @@
 import 'dart:async';
 
-import 'package:aqueduct/src/commands/scripts/get_schema.dart';
+import 'package:aqueduct/src/cli/command.dart';
+import 'package:aqueduct/src/cli/mixins/database_managing.dart';
+import 'package:aqueduct/src/cli/scripts/get_schema.dart';
+import 'package:aqueduct/src/db/schema/schema.dart';
 import 'package:isolate_executor/isolate_executor.dart';
-
-import '../db/db.dart';
-
-import 'base.dart';
-import 'db.dart';
 
 class CLIDatabaseValidate extends CLIDatabaseManagingCommand {
   @override
