+++ aqueduct/lib/src/cli/commands/document_serve.dart
@@ -2,12 +2,12 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:aqueduct/src/commands/running_process.dart';
+import 'package:aqueduct/src/cli/mixins/openapi_options.dart';
+import 'package:aqueduct/src/cli/mixins/project.dart';
+import 'package:aqueduct/src/cli/running_process.dart';
 
-import '../http/http.dart';
-
-import 'base.dart';
-import 'document.dart';
+import 'package:aqueduct/src/cli/command.dart';
+import 'package:aqueduct/src/http/http.dart';
 
 /// Used internally.
 class CLIDocumentServe extends CLICommand with CLIProject, CLIDocumentOptions {
