+++ aqueduct/lib/src/cli/commands/serve.dart
@@ -2,11 +2,12 @@ import 'dart:async';
 import 'dart:io';
 import 'dart:isolate';
 
-import 'package:aqueduct/src/commands/running_process.dart';
-import 'package:aqueduct/src/commands/scripts/get_channel_type.dart';
+import 'package:aqueduct/src/cli/mixins/project.dart';
+import 'package:aqueduct/src/cli/running_process.dart';
+import 'package:aqueduct/src/cli/scripts/get_channel_type.dart';
 import 'package:isolate_executor/isolate_executor.dart';
 
-import 'base.dart';
+import 'package:aqueduct/src/cli/command.dart';
 import 'dart:developer';
 
 class CLIServer extends CLICommand with CLIProject {
