+++ aqueduct/lib/src/cli/mixins/database_managing.dart
@@ -1,0 +1,63 @@
+import 'dart:async';
+import 'dart:io';
+
+import 'package:aqueduct/src/cli/command.dart';
+import 'package:aqueduct/src/cli/mixins/project.dart';
+import 'package:aqueduct/src/cli/scripts/schema_builder.dart';
+import 'package:aqueduct/src/db/schema/migration_source.dart';
+import 'package:aqueduct/src/db/schema/schema.dart';
+import 'package:isolate_executor/isolate_executor.dart';
+
+abstract class CLIDatabaseManagingCommand extends CLICommand with CLIProject {
+  CLIDatabaseManagingCommand() {
+    options
+      ..addOption("migration-directory",
+        help:
+        "The directory where migration files are stored. Relative paths are relative to the application-directory.",
+        defaultsTo: "migrations");
+  }
+
+  Directory get migrationDirectory {
+    final dir = new Directory(decode("migration-directory")).absolute;
+
+    if (!dir.existsSync()) {
+      dir.createSync();
+    }
+    return dir;
+  }
+
+  List<MigrationSource> get projectMigrations {
+    try {
+      final pattern = new RegExp(r"^[0-9]+[_a-zA-Z0-9]*\.migration\.dart$");
+      final sources = migrationDirectory
+        .listSync()
+        .where((fse) => fse is File && pattern.hasMatch(fse.uri.pathSegments.last))
+        .map((fse) => new MigrationSource.fromFile(fse.uri))
+        .toList();
+
+      sources.sort((s1, s2) => s1.versionNumber.compareTo(s2.versionNumber));
+
+      return sources;
+    } on StateError catch (e) {
+      throw new CLIException(e.message);
+    }
+  }
+
+  Future<Schema> schemaByApplyingMigrationSources(List<MigrationSource> sources, {Schema fromSchema}) async {
+    fromSchema ??= new Schema.empty();
+
+    if (sources.isNotEmpty) {
+      displayProgress("Replaying versions: ${sources.map((f)
+      => f.versionNumber.toString()).join(", ")}...");
+    }
+
+    final schemaMap = await IsolateExecutor.executeWithType(SchemaBuilderExecutable,
+      packageConfigURI: packageConfigUri,
+      imports: SchemaBuilderExecutable.imports,
+      additionalContents: MigrationSource.combine(sources),
+      message: SchemaBuilderExecutable.createMessage(sources, fromSchema),
+      logHandler: displayProgress);
+
+    return new Schema.fromMap(schemaMap as Map<String, dynamic>);
+  }
+}
