+++ aqueduct/lib/src/cli/scripts/get_schema.dart
