+++ aqueduct/lib/src/cli/starter.dart
@@ -3,6 +3,10 @@ import 'dart:isolate';
 
 import 'package:aqueduct/aqueduct.dart';
 
+/*
+  Warning: do not remove. This method is invoked by a generated script.
+
+ */
 Future startApplication<T extends ApplicationChannel>(Application<T> app, int isolateCount, SendPort parentPort) async {
   final port = new ReceivePort();
