--- aqueduct/lib/src/executable.dart
@@ -1,3 +1,0 @@
-library aqueduct.executable;
-
-export 'commands/base.dart';
